elimado la parte de guest, que no es necesaria
