+++ src/component/navBar/navBarLateral.js
@@ -84,17 +84,17 @@ export const NavBarLateral = () => {
                 <BtnNav name={"Bookings"} />
             </Div>
 
-            <Div>
+           {/*  <Div>
                 <IconDiv>
                     <FaRegUser/>
                 </IconDiv>                    
                 
                 <BtnNav name={"Guest"} />                
-            </Div>
+            </Div> */}
             
             <Div>
                 <IconDiv>
-                    <RiKey2Line/>
+                    <FaRegUser/>
                 </IconDiv>
 
                 <BtnNav name={"Concierge"} />
