+++ src/index.css
@@ -1,4 +1,5 @@
 body{
     background: #8080800d;
     margin: 0;
+    font-family:sans-serif;
 }
